+++ CHANGELOG.md
@@ -10,6 +10,9 @@
       sudo apt install libqt5quick5 libqt5qml5 qt6-declarative-dev
     ```
 
+- Improved: Error handling in the `install.sh` script
+    - Thank you `@moukhtar22` for finding this and filing an issue
+
 ## 11 April 2026
 
 - Fixed `ags.sh` install script